+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`1.18.4...main`][1.18.4...main].
 
+### Changed
+
+- Allowed installation on PHP 8.6 ([#412]), by [@localheinz]
+
 ## [`1.18.4`][1.18.4]
 
 For a full diff see [`1.18.3...1.18.4`][1.18.3...1.18.4].
@@ -509,5 +513,6 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [#361]: https://github.com/ergebnis/rector-rules/pull/361
 [#408]: https://github.com/ergebnis/rector-rules/pull/408
 [#410]: https://github.com/ergebnis/rector-rules/pull/410
+[#412]: https://github.com/ergebnis/rector-rules/pull/412
 
 [@localheinz]: https://github.com/localheinz