+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`1.18.4...main`][1.18.4...main].
+For a full diff see [`1.18.5...main`][1.18.5...main].
+
+## [`1.18.5`][1.18.5]
+
+For a full diff see [`1.18.4...1.18.5`][1.18.4...1.18.5].
 
 ### Changed
 
@@ -462,7 +466,9 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [1.18.3]: https://github.com/ergebnis/rector-rules/releases/tag/1.18.3
 [1.18.3...1.18.4]: https://github.com/ergebnis/rector-rules/compare/1.18.3...1.18.4
 [1.18.4]: https://github.com/ergebnis/rector-rules/releases/tag/1.18.4
-[1.18.4...main]: https://github.com/ergebnis/rector-rules/compare/1.18.4...main
+[1.18.4...1.18.5]: https://github.com/ergebnis/rector-rules/compare/1.18.4...1.18.5
+[1.18.5]: https://github.com/ergebnis/rector-rules/releases/tag/1.18.5
+[1.18.5...main]: https://github.com/ergebnis/rector-rules/compare/1.18.5...main
 
 [#1]: https://github.com/ergebnis/rector-rules/pull/1
 [#3]: https://github.com/ergebnis/rector-rules/pull/3